Dia 22.05

+++ app/src/main/java/br/senai/rn/agenda/activities/ListaAlunosActivity.java
@@ -5,14 +5,15 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v7.app.AppCompatActivity;
+import android.view.ContextMenu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 
-
-
 import br.senai.rn.activities.R;
+
 import br.senai.rn.agenda.dao.AlunoDAO;
 import br.senai.rn.agenda.models.Aluno;
 
@@ -22,7 +23,7 @@ public class ListaAlunosActivity extends AppCompatActivity {
     private ListView listaAlunos;
     private FloatingActionButton botaoAdicionar;
     private ArrayAdapter<Aluno> adapter;
-    private AlunoDAO dao = new AlunoDAO();
+    private AlunoDAO dao;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -36,16 +37,24 @@ public class ListaAlunosActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
-        adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, dao.obterTodos());
-        listaAlunos.setAdapter(adapter);
+        adapter.clear();
+        adapter.addAll(dao.obterTodos());
     }
 
     private void inicializarComponentes() {
+        dao = new AlunoDAO();
+
         listaAlunos = findViewById(R.id.activity_main_lista_de_alunos);
         botaoAdicionar = findViewById(R.id.activity_main_fab_novo_aluno);
 
-        dao.salvar(new Aluno("Maria","988554477","maria@gmail.com"));
-        dao.salvar(new Aluno("João","988664311","joao@gmail.com"));
+        adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1);
+        listaAlunos.setAdapter(adapter);
+
+        registerForContextMenu(listaAlunos);
+
+        dao.salvar(new Aluno("José", "112233", "jose@gmail"));
+        dao.salvar(new Aluno("João", "445566", "joao@hotmail"));
+        dao.salvar(new Aluno("Pedro", "778899", "pedro@yahoo"));
     }
 
     private void definirEventos() {
@@ -59,22 +68,33 @@ public class ListaAlunosActivity extends AppCompatActivity {
         listaAlunos.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                Aluno alunoSelecionado = dao.obterTodos().get(position);
+                Aluno alunoSelecionado = (Aluno) parent.getItemAtPosition(position);
                 Intent intent = new Intent(ListaAlunosActivity.this, FormularioAlunoActivity.class);
                 intent.putExtra("aluno", alunoSelecionado);
                 startActivity(intent);
             }
         });
-        listaAlunos.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
-            @Override
-            public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
-                Aluno alunoSelecionado = (Aluno)parent.getItemAtPosition(position);
-                dao.remover(alunoSelecionado);
-                adapter.remove(alunoSelecionado);
-                return true;
+    }
 
-            }
-        });
+    @Override
+    public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
+        super.onCreateContextMenu(menu, v, menuInfo);
+        getMenuInflater().inflate(R.menu.activity_lista_alunos_menu, menu);
+    }
+
+    @Override
+    public boolean onContextItemSelected(MenuItem item) {
+        int menuSelecionado = item.getItemId();
+
+        if (menuSelecionado == R.id.activity_lista_aluno_menu_remover) {
+            AdapterView.AdapterContextMenuInfo menuInfo
+                    = (AdapterView.AdapterContextMenuInfo) item.getMenuInfo();
+            Aluno alunoSelecionado = adapter.getItem(menuInfo.position);
+            dao.remover(alunoSelecionado);
+            adapter.remove(alunoSelecionado);
+        }
+
+        return super.onContextItemSelected(item);
     }
 
 }
